initialization system: createGame moved from factory to own file
