+++ src/lib/stores.ts
@@ -1,5 +1,5 @@
 import { writable, type Writable } from 'svelte/store';
 import type { Game } from './types';
-import { createGame } from './factory';
+import { createGame } from './systems/initialization';
 
 export const game: Writable<Game> = writable(createGame());
